#1212 fix test packaging and update multidex proguard rules
