+++ app/multidex-instrumentation-config.pro
@@ -1,5 +1,5 @@
 # Pre android 5.0 test dependencies must be in the primary dex
 
 -keep class org.hamcrest.Matchers
--keep class org.kiwix.kiwixmobile.tests.** { *; }
+-keep class org.kiwix.kiwixmobile.** { *; }
 -keep class org.mockito.** { *; }
