--- app/src/androidTest/java/org/kiwix/kiwixmobile/activities/package-info.java
@@ -1,18 +1,0 @@
-/**
- * These are very simple activity starting and navigation tests.
- * Their purpose is to ensure that navigation is possible without causing crashes.
- *
- * SplashActivity contains various tests of navigating to other activities, along with starting
- * itself.
- *
- * Each test contains:
- *
- * @Test public void empty() {
- *
- * }
- *
- * This forces JUnit to execute the ActivityRule, and then exit immediately (start and stop the
- * Activity)
- */
-
-package org.kiwix.kiwixmobile.activities;
