+++ app/src/androidTest/java/org/kiwix/kiwixmobile/bookmark/BookmarksActivityTest.java
@@ -1,22 +1,19 @@
-package org.kiwix.kiwixmobile.activities;
+package org.kiwix.kiwixmobile.bookmark;
 
 import android.Manifest;
-import androidx.test.filters.LargeTest;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.rule.ActivityTestRule;
 import androidx.test.rule.GrantPermissionRule;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kiwix.kiwixmobile.search.SearchActivity;
 
-@LargeTest
 @RunWith(AndroidJUnit4.class)
-public class SearchActivityTest {
+public class BookmarksActivityTest {
 
   @Rule
-  public ActivityTestRule<SearchActivity> mActivityTestRule = new ActivityTestRule<>(
-      SearchActivity.class);
+  public ActivityTestRule<BookmarksActivity> mActivityTestRule = new ActivityTestRule<>(
+      BookmarksActivity.class);
   @Rule
   public GrantPermissionRule readPermissionRule =
       GrantPermissionRule.grant(Manifest.permission.READ_EXTERNAL_STORAGE);
@@ -25,7 +22,7 @@ public class SearchActivityTest {
       GrantPermissionRule.grant(Manifest.permission.WRITE_EXTERNAL_STORAGE);
 
   @Test
-  public void SearchActivitySimple() {
+  public void testSimple() {
 
   }
 }
