+++ app/src/androidTest/java/org/kiwix/kiwixmobile/search/SearchActivityTest.java
@@ -1,20 +1,21 @@
-package org.kiwix.kiwixmobile.activities;
+package org.kiwix.kiwixmobile.search;
 
 import android.Manifest;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.filters.LargeTest;
 import androidx.test.rule.ActivityTestRule;
 import androidx.test.rule.GrantPermissionRule;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kiwix.kiwixmobile.bookmark.BookmarksActivity;
 
+@LargeTest
 @RunWith(AndroidJUnit4.class)
-public class BookmarksActivityTest {
+public class SearchActivityTest {
 
   @Rule
-  public ActivityTestRule<BookmarksActivity> mActivityTestRule = new ActivityTestRule<>(
-      BookmarksActivity.class);
+  public ActivityTestRule<SearchActivity> mActivityTestRule = new ActivityTestRule<>(
+      SearchActivity.class);
   @Rule
   public GrantPermissionRule readPermissionRule =
       GrantPermissionRule.grant(Manifest.permission.READ_EXTERNAL_STORAGE);
@@ -23,7 +24,7 @@ public class BookmarksActivityTest {
       GrantPermissionRule.grant(Manifest.permission.WRITE_EXTERNAL_STORAGE);
 
   @Test
-  public void testSimple() {
+  public void SearchActivitySimple() {
 
   }
 }
