+++ app/src/androidTest/java/org/kiwix/kiwixmobile/zim_manager/ZimManageRobot.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.manage
+package org.kiwix.kiwixmobile.zim_manager
 
 import applyWithViewHierarchyPrinting
 import org.kiwix.kiwixmobile.BaseRobot
